+++ source/B2B.Transactions/Infrastructure/Authentication/Bearer/Errors/AuthenticationHeaderIsNotBearerToken.cs
@@ -12,19 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Security.Claims;
-
-namespace B2B.Transactions.Infrastructure.Authentication
+namespace B2B.Transactions.Infrastructure.Authentication.Bearer.Errors
 {
-    public class CurrentClaimsPrincipal
+    public class AuthenticationHeaderIsNotBearerToken : AuthenticationError
     {
-        private ClaimsPrincipal? _currentUser;
-
-        public ClaimsPrincipal? ClaimsPrincipal => _currentUser;
-
-        public void SetCurrentUser(ClaimsPrincipal currentUser)
+        public AuthenticationHeaderIsNotBearerToken()
+        : base("The value defined in authorization header is not start with 'bearer'.")
         {
-            _currentUser = currentUser;
         }
     }
 }
