+++ source/B2B.Transactions/Infrastructure/Authentication/Bearer/Errors/TokenValidationFailed.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.Transactions.Infrastructure.Authentication.Errors
+namespace B2B.Transactions.Infrastructure.Authentication.Bearer.Errors
 {
     public class TokenValidationFailed : AuthenticationError
     {
